LayoutTests/webrtc/h265-expected.txt

Video should be running, go to black and running.
Following, should be a snapshot of the video, a black frame and a snapshot of the video.


PASS H265 in getCapabilities
PASS Verify H265 activation
PASS Setting video exchange
PASS Ensuring connection state is connected
PASS Track is enabled, video should not be black
PASS Track is disabled, video should be black
PASS Track is enabled, video should not be black 2
